Added header file defining version number set by build system
With this patch the current version will be shown in splash screen on
application start.

The version number shows:
* Major version number
* Minor version number
* Patch version number, containing:
  - Unique build number (if using a build server) OR
  - Unique git hash (e.g. dev_<shorten git hash> to identify a git commit)

This makes a single version unique based on a build, which is based on a
specific git commit or by identifing a git commit itself if no build
server is used.

Since every patch results in a new commit (and propably in a new build),
every patch/commit results in a unique version.