Create new QRCodeWidget for fullscreen mode
This will create a new QRCodeWidget if fullscreen mode is requested and
will remove this widget if fullscreen mode is not needed any more.

To prevent multiple creation of QRCode widgets set nullptr as parent and
check this value on second object creation.

Not a very nice solution ... needs some review.